bcc: Fixes to allow libbpf to be de-vendored
Source: meta-clang
MR: 121046
Type: Integration
Disposition: Merged from meta-clang
ChangeID: 3ef1d5c79985263436b3a240ea06375cbef7fcf3
Description:

De-vendoring libbpf resolves a SEGV everytime we run bpftrace.
Unfortunately it is not currently possible to de-vendor libbpf because
bcc does not build.

Fix this with brute force and ignorance... and in a seperate patch so
this can be treated as an explanation of the problem and ripped out in
favour of the right fix.

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>
Signed-off-by: Jeremy A. Puhlman <jpuhlman@mvista.com>